+++ activemodel/CHANGELOG.md
@@ -1,2 +1,8 @@
+*   Removed deprecated methods in `ActiveModel::Errors`.
+
+    `#get`, `#set`, `[]=`, `add_on_empty` and `add_on_blank`.
+
+    *Rafael Mendonça França*
+
 
 Please check [5-0-stable](https://github.com/rails/rails/blob/5-0-stable/activemodel/CHANGELOG.md) for previous changes.